// my-app/src/app/utils/types/problemStructure.ts
export type Example = {
	id: number;
	inputText: string;
	outputText: string;
	explanation?: string;
	img?: string;
};

// local problem data
export type Problem = {
	id: string;
	title: string;
	problemStatement: string;
	examples: Example[];
	constraints: string;
	order: number;
	starterCode: string;
	handlerFunction: ((fn: any) => boolean) | string;
	starterFunctionName: string;
    difficulty: string;
};

export type DBProblem = {
	id: string;
	title: string;
	category: string;
	difficulty: string;
	order: number;
	solution?: string;
};